+++ amino.go
@@ -1,7 +1,7 @@
 package crypto
 
 import (
-	amino "github.com/tendermint/go-amino"
+	amino "github.com/teragrid/go-amino"
 )
 
 var cdc = amino.NewCodec()
@@ -18,19 +18,19 @@ func init() {
 func RegisterAmino(cdc *amino.Codec) {
 	cdc.RegisterInterface((*PubKey)(nil), nil)
 	cdc.RegisterConcrete(PubKeyEd25519{},
-		"tendermint/PubKeyEd25519", nil)
+		"teragrid/PubKeyEd25519", nil)
 	cdc.RegisterConcrete(PubKeySecp256k1{},
-		"tendermint/PubKeySecp256k1", nil)
+		"teragrid/PubKeySecp256k1", nil)
 
 	cdc.RegisterInterface((*PrivKey)(nil), nil)
 	cdc.RegisterConcrete(PrivKeyEd25519{},
-		"tendermint/PrivKeyEd25519", nil)
+		"teragrid/PrivKeyEd25519", nil)
 	cdc.RegisterConcrete(PrivKeySecp256k1{},
-		"tendermint/PrivKeySecp256k1", nil)
+		"teragrid/PrivKeySecp256k1", nil)
 
 	cdc.RegisterInterface((*Signature)(nil), nil)
 	cdc.RegisterConcrete(SignatureEd25519{},
-		"tendermint/SignatureKeyEd25519", nil)
+		"teragrid/SignatureKeyEd25519", nil)
 	cdc.RegisterConcrete(SignatureSecp256k1{},
-		"tendermint/SignatureKeySecp256k1", nil)
+		"teragrid/SignatureKeySecp256k1", nil)
 }