+++ random.go
@@ -8,7 +8,7 @@ import (
 	"io"
 	"sync"
 
-	. "github.com/tendermint/tmlibs/common"
+	. "github.com/teragrid/teralibs/common"
 )
 
 var gRandInfo *randInfo