+++ CHANGELOG.md
@@ -16,7 +16,7 @@ BREAKING CHANGES
 
 BREAKING CHANGES
 
-- Update Amino names from "com.tendermint/..." to "tendermint/"
+- Update Amino names from "com.teragrid/..." to "teragrid/"
 
 ## 0.5.0 (March 2, 2018)
 
@@ -98,7 +98,7 @@ BREAKING CHANGES:
 
 FEATURES:
 
-- `github.com/tendermint/go-keys -> github.com/tendermint/go-crypto/keys` - command and lib for generating and managing encrypted keys
+- `github.com/teragrid/go-keys -> github.com/teragrid/go-crypto/keys` - command and lib for generating and managing encrypted keys
 - [hd] New function `WIFFromPrivKeyBytes(privKeyBytes []byte, compress bool) string`
 - Changes to `PrivKey`, `PubKey`, and `Signature` (denoted `Xxx` below):
   - Expose a new method `Unwrap() XxxInner` on the `Xxx` struct which returns the corresponding `XxxInner` interface
@@ -106,7 +106,7 @@ FEATURES:
 
 IMPROVEMENTS:
 
-- Update to use new `tmlibs` repository
+- Update to use new `teralibs` repository
 
 ## 0.1.0 (April 14, 2017)